backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.24.3
google-generativeai==0.3.2
Werkzeug==2.3.7